requirements.txt
# Web Framework
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2

# API Client
anthropic==0.8.1

# Environment Variables
python-dotenv==1.0.0

# Web Server
gunicorn==21.2.0

# Utilities
requests==2.31.0
urllib3==2.0.4
certifi==2023.7.22
idna==3.4
chardet==5.1.0


# Audio Processing
faster-whisper==0.9.0
edge-tts==6.1.9
langdetect==1.0.9




# app 1
yt-dlp
openai-whisper 



# Bill Analyzer
PyMuPDF==1.23.8
pytesseract==0.3.10
Pillow==10.0.1
gunicorn==21.2.0



# CV Analyzer
PyPDF2==3.0.1
python-docx==0.8.11
httpx==0.25.0